+++ R/app_server.R
@@ -11,7 +11,19 @@ app_server <- function(input, output, session) {
   irish_data <- hamiltonCovid19::irish_data
   irish_county_data <- hamiltonCovid19::irish_county_data
 
-  tab_last_updated <- hamiltonCovid19::tab_last_updated
+  global_data <- hamiltonCovid19::global_data
+  # tab_last_updated <- hamiltonCovid19::tab_last_updated
 
-  callModule(mod_summary_ireland_server, "summary_ireland_ui_1")
+  callModule(
+    mod_summary_ireland_server,
+    "summary_ireland_ui_1",
+    irish_data,
+    irish_county_data
+  )
+
+  callModule(
+    mod_summary_global_server,
+    "summary_global_ui_1",
+    global_data
+  )
 }