+++ R/app_ui.R
@@ -85,7 +85,8 @@ create_body <- function() {
     bs4Dash::bs4TabItems(
       bs4Dash::bs4TabItem(
         tabName = "summary",
-        mod_summary_ireland_ui("summary_ireland_ui_1")
+        mod_summary_ireland_ui("summary_ireland_ui_1"),
+        mod_summary_global_ui("summary_global_ui_1")
       )
     )
   )