+++ R/utils-ui.R
@@ -21,8 +21,18 @@ col_8 <- function(...){
 }
 
 #' @importFrom shiny column
+col_7 <- function(...){
+  column(7, ...)
+}
+
+#' @importFrom shiny column
 col_6 <- function(...){
   column(6, ...)
+}
+
+#' @importFrom shiny column
+col_5 <- function(...){
+  column(5, ...)
 }