+++ data-raw/global_data.R
@@ -1,0 +1,44 @@
+## code to prepare `data_global` dataset goes here
+
+library(dplyr)
+
+global_data_raw <- readr::read_rds('data-raw/rds/latest_global_data.rds') %>%
+  tibble::as_tibble()
+
+global_data_raw <- global_data_raw %>%
+  rename(Date = dateRep) %>%
+  mutate(
+    countriesAndTerritories = case_when(
+      countriesAndTerritories == 'Cases_on_an_international_conveyance_Japan' ~
+        'Cruise_ship',
+      countriesAndTerritories == 'United_States_of_America' ~'USA',
+      countriesAndTerritories == 'United_Kingdom' ~ 'UK',
+      TRUE ~ countriesAndTerritories
+    )
+  )
+
+global_data_totals <- global_data_raw %>%
+  mutate(popData2019 = as.numeric(popData2019)) %>%
+  group_by(Date) %>%
+  summarise(
+    across(c("deaths", "cases", "popData2019"), sum, na.rm = TRUE),
+    across(c("day", "month", "year"), first)
+  ) %>%
+  mutate(
+    countriesAndTerritories = 'Global'
+  )
+
+global_data <- bind_rows(global_data_raw, global_data_totals)
+
+global_data <- global_data %>%
+  group_by(countriesAndTerritories) %>%
+  arrange(Date, .by_group = TRUE) %>%
+  mutate(
+    changeCases = cases - lag(cases),
+    changeDeaths = deaths - lag(deaths),
+    totalCases = cumsum(cases),
+    totalDeaths = cumsum(deaths)
+  ) %>%
+  ungroup()
+
+usethis::use_data(global_data, overwrite = TRUE)